chore(core): add hslToRgb utility to exports
