+++ packages/core/src/internal/utils/color.ts
@@ -1,0 +1,27 @@
+export function hslToRgb(h: number, s: number, l: number) {
+  s = s / 100;
+  l = l / 100;
+  const a = s * Math.min(l, 1 - l);
+  const f = (n: any, k = (n + h / 30) % 12) => l - a * Math.max(Math.min(k - 3, 9 - k, 1), -1);
+  return [Math.round(f(0) * 255), Math.round(f(8) * 255), Math.round(f(4) * 255)];
+}
+
+export function rgbToHex(red: number, green: number, blue: number) {
+  let r = red.toString(16);
+  let g = green.toString(16);
+  let b = blue.toString(16);
+
+  if (r.length === 1) {
+    r = '0' + r;
+  }
+
+  if (g.length === 1) {
+    g = '0' + g;
+  }
+
+  if (b.length === 1) {
+    b = '0' + b;
+  }
+
+  return `#${r}${g}${b}`;
+}
