+++ packages/core/src/styles/tokens/tokens.stories.ts
@@ -11,6 +11,7 @@ import { homeIcon } from '@cds/core/icon/shapes/home.js';
 import { plusIcon } from '@cds/core/icon/shapes/plus.js';
 import { trashIcon } from '@cds/core/icon/shapes/trash.js';
 import { downloadIcon } from '@cds/core/icon/shapes/download.js';
+import { hslToRgb, rgbToHex } from '@cds/core/internal';
 
 import '@cds/core/navigation/register.js';
 
@@ -171,34 +172,6 @@ function createToken(token: Token, globals: any, currentTheme: any) {
 
 function tokenToCssProp(name: string) {
   return `--cds-${name.replace(/([A-Z]|\d+)/g, '-$1').toLocaleLowerCase()}`;
-}
-
-function hslToRgb(h: number, s: number, l: number) {
-  s = s / 100;
-  l = l / 100;
-  const a = s * Math.min(l, 1 - l);
-  const f = (n: any, k = (n + h / 30) % 12) => l - a * Math.max(Math.min(k - 3, 9 - k, 1), -1);
-  return [Math.round(f(0) * 255), Math.round(f(8) * 255), Math.round(f(4) * 255)];
-}
-
-function rgbToHex(red: number, green: number, blue: number) {
-  let r = red.toString(16);
-  let g = green.toString(16);
-  let b = blue.toString(16);
-
-  if (r.length === 1) {
-    r = '0' + r;
-  }
-
-  if (g.length === 1) {
-    g = '0' + g;
-  }
-
-  if (b.length === 1) {
-    b = '0' + b;
-  }
-
-  return `#${r}${g}${b}`;
 }
 
 function isAnimationToken(token: Token) {
